producer/requirements.txt
confluent-kafka==2.3.0
websocket-client==1.7.0
requests==2.31.0
python-dotenv==1.0.0
structlog==23.2.0
